merge to master from develop

+++ src/components/clubs/ClubCommunityItem.vue
@@ -1,9 +1,11 @@
 <template>
-  <div>
-    <div class="list text-left">
+  <div class="row d-flex justify-content-center my-5">
+    <div class="col-lg-6 col-md-9 col-11 text-left">
       <div class="items p-2 bg-white px-3 border mt-2">
-        <div class="d-flex justify-content-between align-items-center user-info">
-          <div class="d-flex flex-row align-items-center"><img src="https://i.imgur.com/SzqskfC.png" width="60">
+        <div class="m-2 d-flex justify-content-between align-items-center user-info">
+          <div class="d-flex flex-row align-items-center">
+            <i class="fab fa-instagram fa-3x"></i>
+            <!-- <img src="https://i.imgur.com/SzqskfC.png" width="60"> -->
             <div class="d-flex flex-column user-name-followers ml-3"><span class="font-weight-bold">{{ username }}</span><span>{{ createTime }}  <i class="fas fa-globe-americas"></i></span></div>
           </div>
           <!-- <div><i class="fa fa-bookmark-o fa-2x text-info"></i></div> -->
@@ -13,18 +15,18 @@
           </div>
         </div>
         <div>
-          <h4 class="text-primary">"{{ article.title }}"</h4>
-          <p class="about">{{ article.content }}</p>
-          <img width="100%" src="https://article.images.consumerreports.org/f_auto/prod/content/dam/CRO-Images-2020/Magazine/05May/CR-Health-Inlinehero-HealthyFastFood-3-20-v2" alt="article_img">
-          <!-- <div class="d-flex flex-row mb-3"><span class="px-2 bg-primary text-white mr-1 rounded tags">Football</span><span class="px-2 bg-primary text-white mr-1 rounded tags">Players</span><span class="px-2 bg-primary text-white mr-1 rounded tags">Fans</span></div> -->
-          <hr>
-          <div class="d-flex justify-content-around cursor">
-            <div @click="showCommentsChange"><i class="far fa-comment-alt"> Comment</i></div>
-            <div><i class="fas fa-share"> Share</i></div>
+          <img width="100%" v-if="article.image" :src="ArticleImg" class="p-3" alt="article_img">
+          <div class="mx-3">
+            <h5 class="font-weight-bold">{{ article.title }}</h5>
+            <p class="about">{{ article.content }}</p>
           </div>
           <hr>
+          <div class="ml-3 mb-3 cursor">
+            <div @click="showCommentsChange"><i class="far fa-comment-alt"> Comment</i></div>
+          </div>
+          <!-- <hr> -->
         </div>
-        <CommentList class="mr-3" v-show="showComments" :comments="comments" :commentsNum="commentsNum" :articleId="article.id"/>
+        <CommentList class="mx-3" v-show="showComments" :comments="comments" :commentsNum="commentsNum" :articleId="article.id"/>
       </div>
     </div>
   </div>
@@ -59,6 +61,9 @@ export default {
     createTime(){
       return this.article.created_at.slice(0,10)
     },
+    ArticleImg(){
+      return BACK_URL + this.article.image
+    }
   },
   methods: {
     getComments(){
@@ -73,15 +78,28 @@ export default {
       this.showComments = !this.showComments
     },
     updateArticle(){
-      // router push
+      this.$router.push({ name: 'ArticleUpdateView', params: { 'articleId': this.article.id } })
     },
     deleteArticle(){
-      axios.delete(BACK_URL + `/community/articles/${this.article.id}/`, { headers: { Authorization: `Token ${this.$cookies.get("auth-token")}` }})
-        .then((res) => {
-          console.log(res.data)
-          this.$emit('article-delete', this.article.id)
-        })
-        .catch((err) => { console.log(err.response.data) })
+      this.$confirm(
+        {
+          message: `삭제하시겠습니까?`,
+          button: {
+            yes: 'Yes',
+            no: 'No',
+          },
+          callback: confirm => {
+            if (confirm) {
+              axios.delete(BACK_URL + `/community/articles/${this.article.id}/`, { headers: { Authorization: `Token ${this.$cookies.get("auth-token")}` }})
+                .then((res) => {
+                  console.log(res.data)
+                  this.$emit('article-delete', this.article.id)
+                })
+                .catch((err) => { console.log(err.response.data) })
+            }
+          }
+        }
+      )    
     },
   },
   mounted(){
@@ -91,36 +109,30 @@ export default {
 </script>
 
 <style>
-  .list {
-    width: 50%;
-    margin: 0 auto;
-  }
   .items {
       border-radius: 18px
   }
-
   .tags {
       font-size: 13px;
       line-height: 20px;
       cursor: pointer
   }
-
   .about {
       font-size: 14px
   }
-
   .profile-comment {
     width: 100%;
     height: 100%;
   }
-
   .comment {
     background: #F4F4F4;
     border: none;
     border-radius: 10px;
   }
-
   .cursor {
     cursor: pointer;
+  }
+  .fa-instagram {
+    color: rgb(142, 155, 233);
   }
 </style>
